requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
Office365-REST-Python-Client>=2.5.0
python-dotenv>=1.0.0
streamlit-authenticator>=0.2.0
pyyaml>=6.0
